Renamed AssignmentStatus and AssignmentStatusStamp
