--- impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/experiment/assigment_status/AssigmentStatusRepository.java
@@ -1,7 +1,0 @@
-package uk.ac.ebi.impc_prod_tracker.data.experiment.assigment_status;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface AssigmentStatusRepository extends CrudRepository<AssigmentStatus, Long>
-{
-}
