--- impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/experiment/assigment_status_stamp/AssigmentStatusStampRepository.java
@@ -1,7 +1,0 @@
-package uk.ac.ebi.impc_prod_tracker.data.experiment.assigment_status_stamp;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface AssigmentStatusStampRepository extends CrudRepository<AssigmentStatusStamp, Long>
-{
-}
