+++ impc_prod_tracker/src/main/java/uk/ac/ebi/impc_prod_tracker/data/experiment/assignment_status/AssignmentStatusRepository.java
@@ -1,0 +1,7 @@
+package uk.ac.ebi.impc_prod_tracker.data.experiment.assignment_status;
+
+import org.springframework.data.repository.CrudRepository;
+
+public interface AssignmentStatusRepository extends CrudRepository<AssignmentStatus, Long>
+{
+}
